+++ src/e2e/protractor.conf.js
@@ -34,6 +34,7 @@ exports.config = {
     print: function() {}
   },
   onPrepare() {
+
     require("ts-node").register({
       project: require("path").join(__dirname, "./tsconfig.json")
     });